chore: rename ttl controller package
