+++ pkg/controllers/ttl/utils.go
@@ -1,4 +1,4 @@
-package ttlcontroller
+package ttl
 
 import (
 	"context"
