+++ pkg/controllers/ttl/utils_test.go
@@ -1,4 +1,4 @@
-package ttlcontroller
+package ttl
 
 import (
 	"testing"
